requirements.txt
streamlit
pandas
numpy
matplotlib
uuid
plotly
streamlit
pandas
plotly
streamlit
openai
google-genai
anthropic
requests
google-genai
reportlab
reportlab
google-genai
twilio
twilio
twilio
reportlab
networkx
google-genai
reportlab
boto3
opencv-python
requests
google-genai
deepface
opencv-python
tensorflow
reportlab
seaborn
pillow
requests
deepface
opencv-python
tf-keras
pillow
PyPDF2
reportlab
audio-recorder-streamlit
google-generativeai
pyaudio
pydub
pyttsx3
SpeechRecognition
sendgrid
cryptography
requests
openpyxl
